Move round_* functions from `shims::x86::sse41` module to `shims::x86`
